+++ pkg/apis/scheduling/v1alpha2/bindrequest_types.go
@@ -40,9 +40,6 @@ type ReceivedGPU struct {
 	// This is the portion size that the pod will receive from each connected GPU device
 	// This is a serialized float that should be written as a decimal point number.
 	Portion string `json:"portion,omitempty"`
-
-	// GPU memory amount for the pod on each GPU device (in Mib)
-	GPUMemory string `json:"gpuMemory,omitempty"`
 }
 
 type ResourceClaimAllocation struct {