+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Added
-- Added GPU_MEMORY_LIMIT and GPU_PORTION env vars for GPU sharing pods
+- Added GPU_PORTION env var for GPU sharing pods
 
 ### Changed
 - Changed RUNAI-VISIBLE-DEVICES key in GPU sharing configmap to NVIDIA_VISIBLE_DEVICES